Refactoring NTC into new folder module. - Fixed some bugs due merge with 'devel' (renamed props)
