+++ src/GridCal/Engine/Simulations/OPF/__init__.py
@@ -7,7 +7,5 @@ from GridCal.Engine.Simulations.OPF.ac_opf_ts import OpfAcTimeSeries
 from GridCal.Engine.Simulations.OPF.opf_results import OptimalPowerFlowResults
 from GridCal.Engine.Simulations.OPF.opf_ts_results import OptimalPowerFlowTimeSeriesResults
 from GridCal.Engine.Simulations.OPF.opf_driver import OptimalPowerFlow, OpfSimple, OptimalPowerFlowOptions
-from GridCal.Engine.Simulations.OPF.opf_ntc_driver import OptimalNetTransferCapacityDriver, OptimalNetTransferCapacityOptions, OpfNTC
-from GridCal.Engine.Simulations.OPF.opf_ntc_ts_driver import OptimalNetTransferCapacityTimeSeriesDriver, OptimalNetTransferCapacityTimeSeriesResults
 from GridCal.Engine.Simulations.OPF.opf_ts_driver import OptimalPowerFlowTimeSeries, OpfSimpleTimeSeries
